+++ src/main/java/com/lokoko/domain/image/repository/ReviewImageRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.lokoko.domain.image.entity.QReviewImage;
 import com.lokoko.domain.like.entity.QReviewLike;
 import com.lokoko.domain.product.domain.entity.QProduct;
 import com.lokoko.domain.review.api.dto.response.MainImageReview;
-import com.lokoko.domain.review.entity.QReview;
+import com.lokoko.domain.review.domain.entity.QReview;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
