+++ src/main/java/com/lokoko/domain/video/domain/repository/ReviewVideoRepositoryImpl.java
@@ -4,7 +4,7 @@ import static com.lokoko.domain.like.entity.QReviewLike.reviewLike;
 
 import com.lokoko.domain.product.domain.entity.QProduct;
 import com.lokoko.domain.review.api.dto.response.MainVideoReview;
-import com.lokoko.domain.review.entity.QReview;
+import com.lokoko.domain.review.domain.entity.QReview;
 import com.lokoko.domain.video.domain.entity.QReviewVideo;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
